Refactor in preparation for new selection combinators.  The main change is that the Zipper context can now represent a mapping from the Zipper nodes to /any/ of the nodes in the source hierarchy.  (Previously, the range was restricted to 2nd level nodes).
The only restriction on the mapping is that it must reflect the source hierachy's order.
Formally, if 'f' denotes the mapping, and N1 and N2 are any nodes in the source hierarchy
such that N1 comes before N2 in the source's Preorder.  Then for all top-level zipper nodes
Z1 and Z2:  f(Z1)==N1 and f(Z2)==N2 implies Z1 comes before Z2 in the Zipper's traversal order.